Support for building as subdirectory.
Project can be included and mongoose can be provided separately.
Adding new configuration variables to fine control the build,
excluding the different parts which are not necessary.
Build as library is aggregating parameter to disable the
others(mg, unit tests and examples).
Fixed the certificate to correctly manage if calling from
a different root directory, if the project folder name is different,
THESIS_PROJECT_FOLDER_NAME can be used to set it correctly.